Create kin2csv.py
Added a new python script that converts the output kinematic file to csv, so it can be loaded into QGIS for visualizing the track.

This script is based on plottrack.py and works both on windows and linux. To allow cross platform execution the original "rm" command has been changed to python's "os.remove" function.